graph drawing: Add table.reverse_value_iter(). Use it in NetworkSimplex.
This way, whenever we perform a depth-first search in the network
simplex algorithm (we need to do this many times in fact), nodes will be
pushed on the DFS stack in the order they appear in the TikZ code.

The implication on the GansnerKNV layered algorithm is that the edges
included in the initial spanning tree can be predetermined by writing
down the desired search paths first and add cross/back edges in TikZ
afterwards.